+++ sdk/resourcemanager/maintenance/armmaintenance/updates_client.go
@@ -46,7 +46,7 @@ func NewUpdatesClient(subscriptionID string, credential azcore.TokenCredential, 
 
 // NewListPager - Get updates to resources.
 //
-// Generated from API version 2023-04-01
+// Generated from API version 2023-09-01-preview
 //   - resourceGroupName - Resource group name
 //   - providerName - Resource provider name
 //   - resourceType - Resource type
@@ -102,7 +102,7 @@ func (client *UpdatesClient) listCreateRequest(ctx context.Context, resourceGrou
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01")
+	reqQP.Set("api-version", "2023-09-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -119,7 +119,7 @@ func (client *UpdatesClient) listHandleResponse(resp *http.Response) (UpdatesCli
 
 // NewListParentPager - Get updates to resources.
 //
-// Generated from API version 2023-04-01
+// Generated from API version 2023-09-01-preview
 //   - resourceGroupName - Resource group name
 //   - providerName - Resource provider name
 //   - resourceParentType - Resource parent type
@@ -185,7 +185,7 @@ func (client *UpdatesClient) listParentCreateRequest(ctx context.Context, resour
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-04-01")
+	reqQP.Set("api-version", "2023-09-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil